backend_clean/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pandas>=2.2.0
requests>=2.32.0
python-dotenv>=1.0.0
tls-client>=1.0.0
gnews>=0.3.7
beautifulsoup4>=4.12.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
schedule>=1.2.0
